+++ sketch/pulse.js
@@ -1,10 +1,10 @@
 class Pulse extends Program {
-  constructor(lightMatrix, interval) {
+  constructor(lightMatrix, interval, columnDirection) {
     super(lightMatrix, interval);
 
-    this.currentRowIndex = 0;
-
+    this.currentIndex = 0;
     this.fadeTime = 370;
+    this.pulseAsColumn = columnDirection;
   }
 
   update() {
@@ -15,7 +15,7 @@ class Pulse extends Program {
 
       if (this._isTimeForUpdate()) {
         this._fadeCurrent();
-        this._nextRow();
+        this._next();
         this._lightUpCurrent();
         this.lastChange = millis();
       }
@@ -23,25 +23,35 @@ class Pulse extends Program {
   }
 
   _fadeCurrent() {
-    this._fadeRow(this.currentRowIndex, this.fadeTime);
+    if (this.pulseAsColumn) {
+      this._fadeCol(this.currentIndex, this.fadeTime);
+    } else {
+      this._fadeRow(this.currentIndex, this.fadeTime);
+    }
   }
 
   _lightUpCurrent() {
-    this._lightUpRow(this.currentRowIndex, this.defaultLightColor);
-  }
-
-  _nextRow() {
-    this.currentRowIndex++;
-    if (this.currentRowIndex >= this.lights.rows) {
-      this.currentRowIndex = 0;
+    if (this.pulseAsColumn) {
+      this._lightUpCol(this.currentIndex, this.defaultLightColor);
+    } else {
+      this._lightUpRow(this.currentIndex, this.defaultLightColor);
     }
   }
 
-  _previousRow() {
-    this.currentRowIndex--;
-    if (this.currentRowIndex < 0) {
-      this.currentRowIndex = this.lights.rows -1;
+  _next() {
+    this.currentIndex++;
+    if (this.currentIndex >= this.lights.rows) {
+      this.currentIndex = 0;
     }
   }
+
+  _previous() {
+    this.currentIndex--;
+    if (this.currentIndex < 0) {
+      this.currentIndex = this.lights.rows -1;
+    }
+  }
+
+
 
 }