+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Lights handle their own fading.
 - Light coordination separated into programs, e.g. Pulse, Reverse Pulse.
 - Lights organised as matrix instead of array.
+- Matrix pulse both horizontal and vertical.
 
 ### Changed
 - Project name: Light Array -> Lights. To accommodate matrix configuration.